Upstream maintenance & make six API functions public (#24)
* cargo sort (sorts dependencies)

* Fix six doc comments

* cargo fmt

* api: Make 6 utility functions public

These are useful for users of the library,
if they have both a graph and separate `Vec`s or `HashMap`s
referencing parts of the graph
(e.g. all subjects of special interest).
To canonicalize in this scenario,
one would want to create the mapping with `issue` on the graph,
and then apply this mapping both to that graph
and to the separate collections of items in the graph.

* Apply clippy fixes (automatically)